+++ lib/td_qx/expressions/expression_values/function_arg.ex
@@ -21,4 +21,14 @@ defmodule TdQx.Expressions.ExpressionValues.FunctionArg do
     |> cast_embed(:expression, with: &Expression.changeset/2)
     |> validate_required([:name])
   end
+
+  def to_json([%__MODULE__{} | _] = args) do
+    args
+    |> Enum.map(fn %{name: name, expression: expression} ->
+      {name, Expression.to_json(expression)}
+    end)
+    |> Enum.into(%{})
+  end
+
+  def to_json(_), do: nil
 end